requirements.txt
asgiref
Django
djangorestframework
psycopg2-binary
pytz
sqlparse
pillow
djangorestframework-simplejwt